backend/requirements.txt
# Flask and related packages
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0

# Audio processing
pydub==0.25.1
SpeechRecognition==3.10.0

# Whisper for transcription
git+https://github.com/openai/whisper.git
torch==2.0.1
numpy==1.24.3

# LiteLLM for Claude integration
litellm==1.12.1
requests==2.31.0

# Utilities
Werkzeug==2.3.7
